feat: Add cluster_id in structures
Not sure if we should add the "duplicates" in the Structure schema
as they will only be present in some, but not all, endpoints and
situations.

For instance, it won't be present in the datagouv exports,
or as the linked Structure of a service in the service endpoints.

Should we split between a "Structure" and "StructureWithDuplicates" ?